+++ src/abstractdataframe/subset.jl
@@ -31,6 +31,23 @@ function _and_missing(x::Any...)
                         "but only true, false, or missing are allowed"))
 end
 
+function assert_bool_vec(@nospecialize(fun), skipmissing)
+    # we are guaranteed that ByRow returns a vector
+    # this workaround is needed for 0-argument ByRow
+    fun isa ByRow && return fun
+
+    return function(x...)
+        val = fun(x...)
+        T = skipmissing ? Union{Missing, Bool} : Bool
+        # this technically allows T to be Union{} but it will be disallowed later
+        if !(val isa AbstractVector{<:T})
+            throw(ArgumentError("functions passed to `subset` must return " *
+                                "AbstractVector{$T}."))
+        end
+        return val
+    end
+end
+
 # Note that _get_subset_conditions will have a large compilation time
 # if more than 32 conditions are passed as `args`.
 function _get_subset_conditions(df::Union{AbstractDataFrame, GroupedDataFrame},
@@ -39,7 +56,7 @@ function _get_subset_conditions(df::Union{AbstractDataFrame, GroupedDataFrame},
     conditions = Any[if a isa ColumnIndex
                          a => Symbol(:x, i)
                      elseif a isa Pair{<:Any, <:Base.Callable}
-                         first(a) => last(a) => Symbol(:x, i)
+                         first(a) => assert_bool_vec(last(a), skipmissing) => Symbol(:x, i)
                      else
                          throw(ArgumentError("condition specifier $a is not supported by `subset`"))
                      end for (i, a) in enumerate(args)]
@@ -77,7 +94,10 @@ Each argument passed in `args` can be either a single column selector or a
 described for [`select`](@ref).
 
 Note that as opposed to [`filter`](@ref) the `subset` function works on whole
-columns (or all rows in groups for `GroupedDataFrame`).
+columns and always makes a subset of rows (or all rows in groups for
+`GroupedDataFrame`) and must return a vector. If you want to filter rows
+of an `AbstractDataFrame` or groups of `GroupedDataFrame` by a function that
+returns `Bool` value use [`filter`](@ref).
 
 If `skipmissing=false` (the default) `args` are required to produce vectors
 containing only `Bool` values. If `skipmissing=true`, additionally `missing` is
@@ -180,7 +200,11 @@ Each argument passed in `args` can be either a single column selector or a
 described for [`select`](@ref).
 
 Note that as opposed to [`filter!`](@ref) the `subset!` function works on whole
-columns (or all rows in groups for `GroupedDataFrame`).
+columns (or all rows in groups for `GroupedDataFrame`) and must return a vector.
+If you want to filter rows of an `AbstractDataFrame` or groups of
+`GroupedDataFrame` by a function that returns `Bool` value use [`filter!`](@ref)
+or [`filter`](@ref) (note that [`filter!`](@ref) is not supported for
+`GroupedDataFrame`).
 
 If `skipmissing=false` (the default) `args` are required to produce vectors
 containing only `Bool` values. If `skipmissing=true`, additionally `missing` is
@@ -191,7 +215,10 @@ If `ungroup=false` the resulting data frame is re-grouped based on the same
 grouping columns as `gdf` and a `GroupedDataFrame` is returned.
 
 If `GroupedDataFrame` is subsetted then it must include all groups present in the
-`parent` data frame, like in [`select!`](@ref).
+`parent` data frame, like in [`select!`](@ref). Also note that in general in this
+case the parent of the passed `GroupedDataFrame` is mutaded in place, which means
+that the source `GroupedDataFrame` is not safe to use as most likely it will
+get corrupted due to row removal in the parent.
 
 See also: [`subset`](@ref), [`filter!`](@ref), [`select!`](@ref)