+++ NEWS.md
@@ -1,3 +1,13 @@
+# DataFrames v1.0.x Patch Release Notes
+
+* patch v1.0.3: make sure `subset` checks if the passed condition function
+  returns a vector of values
+  ([#XXXX](https://github.com/JuliaData/DataFrames.jl/pull/XXXX))
+* patch v1.0.2: fix of perfromance issue of `groupby` when using multi-threading
+  ([#2736](https://github.com/JuliaData/DataFrames.jl/pull/2736))
+* patch v1.0.1: fix of perfromance issue of `groupby` when using `PooledVector`
+  ([2733](https://github.com/JuliaData/DataFrames.jl/pull/2733))
+
 # DataFrames v1.0 Release Notes
 
 ## Breaking changes